+++ src/main/java/dogpath/server/dogpath/global/util/JSONUtils.java
@@ -1,6 +1,5 @@
-package dogpath.server.dogpath.domain.path.algorithm;
+package dogpath.server.dogpath.global.util;
 
-import com.squareup.okhttp.Response;
 import lombok.extern.slf4j.Slf4j;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
